feat: limpando a pasta utils
